Move show_diff to its own file
