+++ s4/cli.py
@@ -5,10 +5,7 @@ import datetime
 import json
 import logging
 import os
-import shutil
-import subprocess
 import sys
-import tempfile
 from collections import defaultdict
 
 import boto3
@@ -23,6 +20,7 @@ from s4 import VERSION
 from s4 import sync
 from s4 import utils
 from s4.clients import local, s3
+from s4.diff import show_diff
 from s4.inotify_recursive import INotifyRecursive
 from s4.progressbar import ProgressBar
 from s4.resolution import Resolution
@@ -61,53 +59,6 @@ def handle_conflict(key, action_1, client_1, action_2, client_2):
         return Resolution.get_resolution(key, action_1, client_2, client_1)
     elif choice == '2':
         return Resolution.get_resolution(key, action_2, client_1, client_2)
-
-
-def show_diff(client_1, client_2, key):
-    if shutil.which("diff") is None:
-        print('Missing required "diff" executable.')
-        print("Install this using your distribution's package manager")
-        return
-
-    if shutil.which("less") is None:
-        print('Missing required "less" executable.')
-        print("Install this using your distribution's package manager")
-        return
-
-    so1 = client_1.get(key)
-    data1 = so1.fp.read()
-    so1.fp.close()
-
-    so2 = client_2.get(key)
-    data2 = so2.fp.read()
-    so2.fp.close()
-
-    fd1, path1 = tempfile.mkstemp()
-    fd2, path2 = tempfile.mkstemp()
-    fd3, path3 = tempfile.mkstemp()
-
-    with open(path1, 'wb') as fp:
-        fp.write(data1)
-    with open(path2, 'wb') as fp:
-        fp.write(data2)
-
-    # This is a lot faster than the difflib found in python
-    with open(path3, 'wb') as fp:
-        subprocess.call([
-            'diff', '-u',
-            '--label', client_1.get_uri(key), path1,
-            '--label', client_2.get_uri(key), path2,
-        ], stdout=fp)
-
-    subprocess.call(['less', path3])
-
-    os.close(fd1)
-    os.close(fd2)
-    os.close(fd3)
-
-    os.remove(path1)
-    os.remove(path2)
-    os.remove(path3)
 
 
 def get_s3_client(target, aws_access_key_id, aws_secret_access_key, region_name):
