+++ tests/test_cli.py
@@ -140,44 +140,6 @@ class TestHandleConflict(object):
         show_diff.assert_called_with(s3_client, local_client, 'movie')
 
 
-class TestShowDiff(object):
-    @mock.patch('shutil.which')
-    def test_diff_not_found(self, which, capsys, local_client, s3_client):
-        which.return_value = None
-        cli.show_diff(local_client, s3_client, "something")
-
-        out, err = capsys.readouterr()
-        assert out == (
-            'Missing required "diff" executable.\n'
-            "Install this using your distribution's package manager\n"
-        )
-
-    @mock.patch('shutil.which')
-    def test_less_not_found(self, which, capsys, local_client, s3_client):
-        def missing_less(value):
-            return None if value == 'less' else 'something'
-
-        which.side_effect = missing_less
-        cli.show_diff(local_client, s3_client, "something")
-
-        out, err = capsys.readouterr()
-        assert out == (
-            'Missing required "less" executable.\n'
-            "Install this using your distribution's package manager\n"
-        )
-
-    @mock.patch('subprocess.call')
-    def test_diff(self, call, local_client, s3_client):
-        utils.set_local_contents(local_client, "something", 4000, "wow")
-        utils.set_s3_contents(s3_client, "something", 3000, "nice")
-
-        cli.show_diff(local_client, s3_client, "something")
-
-        assert call.call_count == 2
-        assert call.call_args_list[0][0][0][0] == "diff"
-        assert call.call_args_list[1][0][0][0] == "less"
-
-
 def test_progressbar_smoketest(s3_client, local_client):
     # Just test that nothing blows up
     utils.set_local_contents(
